navigation and header changes
Fixed some not closed div's and head's and placed the header's after
!body. Anywayyyyyssss, changed the navigation to get dropdown, and some
other changes to the navi, also changed the header 'background img' to
background color................ and placed the logo on top.

In management, the dropdown is overflown by the management panel etc. so
logout and help button don't work under settings, will fix this asap